+++ wayflowcore/tests/mcptools/start_mcp_server.py
@@ -6,13 +6,16 @@
 import argparse
 from contextvars import ContextVar
 from os import PathLike
-from typing import Annotated, Literal, Optional
+from typing import Annotated, AsyncGenerator, Literal, Optional
 
+import anyio
 from mcp.server.fastmcp import FastMCP as BaseFastMCP
 from mcp.types import EmbeddedResource, TextResourceContents
 from pydantic import AnyUrl, BaseModel, Field, RootModel
 from starlette.applications import Starlette
 from typing_extensions import TypedDict
+
+from wayflowcore.mcp.mcphelpers import mcp_streaming_tool
 
 UvicornExtraConfig = TypedDict(
     "UvicornExtraConfig",
@@ -141,6 +144,16 @@ def create_server(host: str, port: int):
             ),
             type="resource",
         )
+
+    @server.tool(description="Streaming tool")
+    @mcp_streaming_tool
+    async def streaming_tool() -> AsyncGenerator[str, None]:
+        contents = [f"This is the sentence N°{i}" for i in range(5)]
+        for chunk in contents:
+            yield chunk  # streamed chunks
+            await anyio.sleep(0.2)
+
+        yield ". ".join(contents)  # final result
 
     return server